Updated by Library Project Script . 20200222

+++ include/mmap-windows.c
@@ -13,7 +13,7 @@
 
 #include "mmap.h"
 
-void *mmap(void *start, size_t length, int prot, int flags, int fd, off_t offset)
+void *mmap(void *start, uint32_t length, int prot, int flags, int fd, off_t offset)
 {
 	if (prot & ~(PROT_READ | PROT_WRITE | PROT_EXEC))
 		return MAP_FAILED;
@@ -64,7 +64,7 @@ void *mmap(void *start, size_t length, int prot, int flags, int fd, off_t offset
 	return ret;
 }
 
-void munmap(void *addr, size_t length)
+void munmap(void *addr, uint32_t length)
 {
 	UnmapViewOfFile(addr);
 }
